Updated GUI and created csv for GUI
GUI is near complete, but without testing. Further logic has been added if the user decides to press any buttons on the right hand panel before initializing the measurement. What is left is to complete the measurement portion of the gui and complete testing. The CSV included is used in order for multiple wavelengths to have different exposures and currently meant for testing